requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
colorama==0.4.6
Django==5.0.6
django-cors-headers==4.7.0
django-environ==0.11.2
django-otp==1.6.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
